More work on warnings

+++ src/Image.cpp
@@ -1025,11 +1025,11 @@ void Image::Recalculate(double scale) {
   if ((m_maxWidth > 0) &&
       (static_cast<wxCoord>(scale) * width >
        m_maxWidth * static_cast<wxCoord>(m_configuration->GetPPI().x)))
-    scale = m_maxWidth * m_configuration->GetPPI().x / width;
+    scale = static_cast<double>(m_maxWidth) * m_configuration->GetPPI().x / width;
   if ((m_maxHeight > 0) &&
       (static_cast<wxCoord>(scale) * height >
        m_maxHeight * static_cast<wxCoord>(m_configuration->GetPPI().y)))
-    scale = m_maxHeight * m_configuration->GetPPI().y / height;
+    scale = static_cast<double>(m_maxHeight) * m_configuration->GetPPI().y / height;
 
   // Set the width of the scaled image
   m_height = static_cast<int>(scale * height);
